refactor(harfbuzz): split HBNewApiTest.cs into appropriate test files
- Moved Language tests to HBCommonTest.cs
- Moved Tag tests to HBCommonTest.cs
- Moved Variation tests to HBCommonTest.cs
- Moved Buffer.CreateSimilar tests to HBBufferTest.cs
- Moved Font tests (SyntheticSlant, SyntheticBold, Variations, NamedInstance) to HBFontTest.cs
- Deleted HBNewApiTest.cs

All 130 HarfBuzz tests pass.
